Conflicts:
	MARS_NB/MARS/MARSCore/src/mars/Initializer.java
	MARS_NB/MARS/MARSCore/src/mars/auv/AUV.java
	MARS_NB/MARS/MARSCore/src/mars/auv/AUV_Manager.java
	MARS_NB/MARS/MARSCore/src/mars/auv/CommunicationManager.java
	MARS_NB/MARS/MARSCore/src/mars/sensors/UnderwaterModem.java
	MARS_NB/MARS/MARSCore/src/mars/sensors/WiFi.java
	MARS_NB/MARS/MARSCore/src/mars/server/Connection.java
	MARS_NB/MARS/MARSCore/src/mars/states/MapState.java
	MARS_NB/MARS/MARSCore/src/mars/states/SimState.java

+++ MARS_NB/MARS/MARSCore/src/mars/sensors/Accelerometer.java
@@ -130,11 +130,11 @@ public class Accelerometer extends Sensor implements ChartValue {
             return getAccelerationRaw();
         } else if (getNoiseType() == NoiseType.UNIFORM_DISTRIBUTION) {
             float noise = getUnifromDistributionNoise(getNoiseValue());
-            Vector3f noised = new Vector3f(getAccelerationRaw().x + ((float) ((1f / 100f) * noise)), getAccelerationRaw().y + ((float) ((1f / 100f) * noise)), getAccelerationRaw().z + ((float) ((1f / 100f) * noise)));
+            Vector3f noised = new Vector3f(getAccelerationRaw().x + (((1f / 100f) * noise)), getAccelerationRaw().y + (((1f / 100f) * noise)), getAccelerationRaw().z + (((1f / 100f) * noise)));
             return noised;
         } else if (getNoiseType() == NoiseType.GAUSSIAN_NOISE_FUNCTION) {
             float noise = getGaussianDistributionNoise(getNoiseValue());
-            Vector3f noised = new Vector3f(getAccelerationRaw().x + ((float) ((1f / 100f) * noise)), getAccelerationRaw().y + ((float) ((1f / 100f) * noise)), getAccelerationRaw().z + ((float) ((1f / 100f) * noise)));
+            Vector3f noised = new Vector3f(getAccelerationRaw().x + (((1f / 100f) * noise)), getAccelerationRaw().y + (((1f / 100f) * noise)), getAccelerationRaw().z + (((1f / 100f) * noise)));
             return noised;
         } else {
             return getAccelerationRaw();
